Fix image color on metadata generation (#33)
* Fix image color on metadata generation

* Pass image color to zora mint and share buttons
